requirements.txt
python-fasthtml
supabase